fix: avoid requesting unnecessary permission when choosing media from the gallery
Context: When using intents or the Photo Picker API for the user to select media (images and videos) we don't need to request the READ_EXTERNAL_STORAGE permission. That is only necessary when accessing files in the external storage (e.g. when calling editURIPicture)

References: https://outsystemsrd.atlassian.net/browse/RMET-3875